player_wearclothes changes, string changes and maintain area moving
Adds 2 toggles for admins to modify:

DZE_backpackRemove forces the player to drop his/her backpack when they
change gear.
DZE_maintainCurrencyRate modifies the amount of worth each item is worth
to maintain.

Changes the error message in player_wearClothes to tell the player why
they can't change gear yet
Remove some debug lines that spam RPT
