+++ SQF/dayz_code/Configs/CfgPlayerStats/p_cover.hpp
