+++ SQF/dayz_code/stringtable.xml
@@ -6885,7 +6885,7 @@
 			<German>Installation abgeschlossen, bitte warten...</German>
 		</Key>
 		<Key ID="str_player_desync">
-			<English>Date and Time not Synced Disconnect.</English>
+			<English>Date and Time not Synced, Disconnect and try again.</English>
 			<German>Datum und Zeit sind nicht synchronisiert. Verbindung wird getrennt!</German>
 			<Russian>Дата и время не синхронизированы. Отключение.</Russian>
 			<Spanish>Fecha y Hora no sincronizada. Desconectando.</Spanish>
@@ -6940,7 +6940,7 @@
 			<German>Charakterdaten vom Server erhalten</German>
 		</Key>
 		<Key ID="str_login_alreadyDead">
-			<English>Player Already Dead, disconnect and try again!</English>
+			<English>Player is already dead, disconnect and try again!</English>
 			<Russian>Игрок уже мертв! Отключитесь и попробуйте снова!</Russian>
 			<French>Vous êtes déjà mort, essayez encore!</French>
 			<Czech>Hráč je mrtev, odhlašte se a zkuste znovu!</Czech>
@@ -7025,7 +7025,6 @@
 			<English>You have finished collecting wood.</English>
 			<Russian>Вы закончили рубку дров.</Russian>
 			<German>Du hast erfolgreich Holz gesammelt.</German>
-			<Spanish>You have finished collecting wood.</Spanish>
 			<French>Vous avez terminé de faire du bois.</French>
 			<Czech>Nasekali jste dřevo.</Czech>
 		</Key>
@@ -7131,7 +7130,6 @@
 			<German>%1 kann nicht auf den Rücken genommen werden.</German>
 		</Key>
 		<Key ID="str_player_31">
-			<!-- %1=ITEM %2=build/consumed/cook/drink/fill/pitch/place/use/wear -->
 			<English>%1 must be in your main inventory to %2 it.</English>
 			<German>%1 muss sich in deinem Inventar befinden, um es zu %2.</German>
 			<Russian>Чтобы %2, (%1) должен находиться в основном инвентаре.</Russian>
@@ -7296,6 +7294,9 @@
 			<Spanish>¡Espera a terminar la acción previa para realizar otra!</Spanish>
 			<French>Attendez la fin de l&apos;action en cours avant d&apos;en commencer une autre!</French>
 			<Czech>Než začnete dělat něco jiného, musíte dokončit aktuální akci.</Czech>
+		</Key>
+		<Key ID="str_player_clotheslimit">
+			<English>You must wait another %1 seconds before attempting to change clothes.</English>
 		</Key>
 		<Key ID="str_player_ammo_successful">
 			<English>You have successfully loaded %1 ammunition.</English>
@@ -13561,14 +13562,14 @@
 			<Czech>Označil si cíl za přítele. Čeká se na přijetí.</Czech>
 		</Key>
 		<Key ID="STR_EPOCH_ACTIONS_9">
-			<English>Unable to change clothes while wearing backpack.</English>
+			<English>You are unable to change clothes while wearing backpack.</English>
 			<German>Kleider können nicht gewechselt werden, solange ein Rucksack getragen wird.</German>
 			<Russian>Нельзя переодеться с рюкзаком на спине.</Russian>
 			<French>Impossible de changer ses vêtements lorsque vous portez un sac à dos.</French>
 			<Czech>Nemůžete si změnit oblečení, pokud nosíte batoh.</Czech>
 		</Key>
 		<Key ID="STR_EPOCH_ACTIONS_10">
-			<English>Unable to change clothes while carrying a filled chainsaw.</English>
+			<English>You are unable to change clothes while carrying a filled chainsaw.</English>
 			<German>Kleider können nicht gewechselt werden, solange eine gefüllte Motorsäge in den Händen gehalten wird.</German>
 			<Russian>Нельзя переодеваться, держа заправленную бензопилу.</Russian>
 			<French>Impossible de changer ses vêtements lorsque vous avez une tronçonneuse en main.</French>
